+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## unreleased
+* Add `isNetworkTokenized` field to `AndroidPayCard` and `AndroidPayDetails`
+
 ## 2.106.0
 * Add `PROCESSOR_DOES_NOT_SUPPORT_MOTO_FOR_CARD_TYPE` to validation errors
 * Update `jackson-jr` to v2.9.9 to fix CVE-2018-11307 (#75 thanks @sehrope)