+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add `TRANSACTION_REFUND_AUTH_HARD_DECLINED` and `TRANSACTION_REFUND_AUTH_SOFT_DECLINED` validation errors
 * Add `isNetworkTokenized` field to `AndroidPayCard` and `AndroidPayDetails`
 * Add GraphQL ID to `CreditCardVerification`, `Customer`, `Dispute`, and `Transaction`
+* Add `retrievalReferenceNumber` field to `Transaction`
 
 ## 2.106.0
 * Add `PROCESSOR_DOES_NOT_SUPPORT_MOTO_FOR_CARD_TYPE` to validation errors